requirements.txt
python-dotenv==1.0.0
openai==1.12.0
twilio==8.13.0
websockets==12.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.6.1
sounddevice==0.4.6
numpy==1.24.3
soundfile==0.12.1
backoff==2.2.1
scipy==1.11.4
